// src/models/user.ts
export interface IUser {
    name: string;
    email: string;
    phone: string;
    GSTIN: string;
    companyName: string;
    address: string;
    bankName: string;
    accountNo: string;
    ifscCode: string;
    branchName: string;
    upiId: string;
    upiQrcode?: string;
    logoImage?: string;
    signImg?: string;
}